- Pass correct dialogs settings to HandleOverlayOnShow - New `CanCloseDialogOwner` property at `MetroDialogSettings` which can be used to handle how the owner of the dialog can be closed. - New `IsCloseButtonEnabledWithDialog` property at `MetroWindow` which indicates if the close button should be enabled or not if a dialog is shown.
